% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_npc.R
\name{calc_npc}
\alias{calc_npc}
\title{nPC Calculation}
\usage{
calc_npc(var, N = 1, cutoff_divisor = 10)
}
\arguments{
\item{var}{a vector; proportion of variation explainded by each of the principal coordinates.}
}
\value{
This function is used internally in the method \code{nPC};
see the help page for \code{nPC} for more details.
}
\description{
calculates nPC.
}
\details{
This function is used internally in the method \code{nPC};
see the help page for \code{nPC} for more details.
}

