% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa.R
\name{cidrPcoa}
\alias{cidrPcoa}
\title{Principal Coordinates Analysis}
\usage{
cidrPcoa(D)
}
\arguments{
\item{D}{a dissimilarity matrix}
}
\value{
This function is used internally in the method \code{scPCA};
see the help page of \code{scPCA} for more details.
}
\description{
Performs principal coordinates analysis on a dissimilarity matrix.
}
\details{
This function is used internally in the method \code{scPCA};
see the help page of \code{scPCA} for more details.

This function is a modified version of the \emph{pcoa} function from the CRAN
package \emph{ape (version 3.5)}.  The function was modified on 6th August 2016.
}
\references{
Paradis E., Claude J. & Strimmer K. 2004. APE: analyses of phylogenetics and evolution in R language. Bioinformatics 20: 289-290.
}

