% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\docType{methods}
\name{determineDropoutCandidates}
\alias{determineDropoutCandidates}
\title{Determine Dropout Candidates}
\usage{
\S4method{determineDropoutCandidates}{scData}(object, min1 = 3, min2 = 8,
  N = 2000, alpha = 0.1, fast = TRUE)
}
\arguments{
\item{object}{the scData class object}

\item{min1}{technical parameter used in estimating the minimum point between the first two modes of the density curve of logTPM for each cell}

\item{min2}{technical parameter used in estimating the minimum point between the first two modes of the density curve of logTPM for each cell}

\item{N}{number of cells to consider when determining the threshold value for dropout candidates; used in conjunction with the \code{fast} parameter}

\item{alpha}{a cutoff quantile in the range [0,1].  Thresholds outside this will be adjusted to the quantile boundary.}

\item{fast}{boolean; if \code{TRUE} (default), implements a fast version for datasets with a sample size greater than N}
}
\value{
an updated scData class object with the following attributes updated

\item{dThreshold}{a vector of library dependent dropout candidate thresholds}
\item{dropoutCandidates}{a matrix with the same dimension as nData. The value is \code{True} for an individual entry if it
is a dropout candidate, otherwise the value is \code{false}.}
}
\description{
determines which entries in a single cell RNA-Seq dataset are potentially dropouts.
}
\details{
populates a truth matrix with the same dimension as the nData.
The value is \code{True} for an individual entry if it
is a dropout candidate, otherwise the value is \code{false}.
}
\examples{
example(cidr)
}

