% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\docType{methods}
\name{nCluster}
\alias{nCluster}
\title{nCluster}
\usage{
\S4method{nCluster}{scData}(object, n = object@nPC * 2 + 2)
}
\arguments{
\item{object}{the scData class object}

\item{n}{maximum number of clusters}
}
\description{
outputs the Calinski-Harabasz Index versus number of clusters plot.
}
\details{
this method outputs the Calinski-Harabasz Index versus number of clusters plot,
useful for deciding the number of clusters.
Since \emph{CIDR} automatically decides the number of clusters,
this method is optional and only used if the user wants to alter the default number of clusters.
}
\examples{
example(cidr)
}

