% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\docType{methods}
\name{nPC}
\alias{nPC}
\title{Determine nPC}
\usage{
\S4method{nPC}{scData}(object)
}
\arguments{
\item{object}{an scData class object.}
}
\value{
an updated scData class object with the following attribute updated

\item{nPC}{the number of principal coordinates to be used in clustering.}
}
\description{
determines the optimal number of principal coordinates (nPC) to be used in clustering.
}
\examples{
example(cidr)
}

