% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\docType{methods}
\name{scCluster}
\alias{scCluster}
\title{Sincle Cell Clustering}
\usage{
\S4method{scCluster}{scData}(object, n = object@nPC * 2 + 2)
}
\arguments{
\item{object}{the scData class object}

\item{n}{Calinski-Harabasz Index is used to decide which number between 2 and n is optimal as the number of clusters.}
}
\value{
an updated scData class object with the following attributes updates

\item{nCluster}{the number of clusters}
\item{clusters}{a vector assigning each cell to a cluster}
}
\description{
performs heirarchical clustering based on \emph{CIDR} principal coordinates.
}
\examples{
example(cidr)
}

