% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\name{scDataConstructor}
\alias{scDataConstructor}
\title{scData Constructor}
\usage{
scDataConstructor(tags)
}
\arguments{
\item{tags}{a matrix of tags where the rows crrespond to features (genes, transcripts, etc) and the columns correspond to cells}
}
\value{
an scData class object
}
\description{
\code{scDataConstructor} creates a new scData class object from a tag table.
}
\details{
Creates an object whose type represents a single-cell RNA-Seq dataset class.
Attributes of the class include scalar, vector and matrix
data types necessary for the CIDR analysis - such as tag table, library
sizes, dropout candidates, imputation weighting threshold.
}
\examples{
## Generate simulated single-cell RNA-Seq tags.
N=3 ## 3 cell types
k=50 ## 50 cells per cell type
sData <- scSimulator(N=N, k=k)
## The input for cidr should be a tag matrix. 
tags <- as.matrix(sData$tags)
## create a new scData object
sData <- scDataConstructor(tags)
## print the first 5 library sizes
sData@librarySizes[1:5]
## print a portion of the data matrix of the class - contains raw tags
sData@tags[1:5, 30:34]
## print part of the data matrix of the class - log tag per million
sData@nData[1:5, 30:34]
}

