% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\docType{methods}
\name{scPCA}
\alias{scPCA}
\title{Single Cell Principal Coordinates Analysis}
\usage{
\S4method{scPCA}{scData}(object)
}
\arguments{
\item{object}{the scData class object}
}
\value{
an updated scData class object with the following attributes updated

\item{eigenvalues}{all eigenvalues (positive and negative) output by the principal coordinates analysis}
\item{PC}{principal coordinates}
\item{variation}{proportion of variation explained by each of the principal coordinates}
\item{nPC}{the number of principal coordinates used in the clustering}
}
\description{
performs principal coordinates analysis on the \emph{CIDR} dissimilarity matrix
}
\examples{
example(cidr)
}

