% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cidrPcoa.R
\name{cidrPcoa}
\alias{cidrPcoa}
\title{Principal Coordinates Analysis}
\usage{
cidrPcoa(D)
}
\arguments{
\item{D}{a dissimilarity matrix.}
}
\value{
This function is used internally in the method \code{scPCA};
see the help page for \code{scPCA} for more details.
}
\description{
performs principal coordinates analysis on a dissimilarity matrix.
}
\details{
This function is used internally in the method \code{scPCA};
see the help page for \code{scPCA} for more details.
}

