% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\docType{methods}
\name{determineDropoutCandidates}
\alias{determineDropoutCandidates}
\title{Determine Dropout Candidates}
\usage{
\S4method{determineDropoutCandidates}{scData}(object, min1 = 3, min2 = 8,
  N = 2000, alpha = 0.1, fast = TRUE, zerosOnly = FALSE,
  bw_adjust = 1)
}
\arguments{
\item{object}{the scData class object.}

\item{min1, min2}{technical parameters used in estimating the minimum point between the first two modes of the density curve of logTPM for each cell.}

\item{N}{number of cells to consider when determining the threshold value for dropout candidates; used in conjunction with the \code{fast} parameter.}

\item{alpha}{a cutoff quantile in the range [0,1]. Thresholds outside this will be adjusted to the quantile boundary.}

\item{fast}{Boolean; if \code{TRUE} (default - unless \code{tagType} is \code{cpm}), implements a fast version for datasets with a sample size greater than N. NOTE: set to \code{FALSE} if \code{tagType} is \code{cpm}.}

\item{zerosOnly}{Boolean; if \code{TRUE}, only zeros are considered as dropout candidates; by default \code{FALSE}.}

\item{bw_adjust}{bandwidth adjustment factor; \emph{CIDR} uses the default bandwidth selection method ‘nrd0’ in the kernel density estimation; see \code{stats::density} help page for more details.}
}
\value{
an updated scData class object with the following attributes updated

\item{dThreshold}{a vector of library dependent dropout candidate thresholds.}
\item{dropoutCandidates}{a matrix with the same dimension as nData. The value is \code{TRUE} for an entry if it
is a dropout candidate; otherwise the value is \code{FALSE}.}
}
\description{
determines which entries in a single-cell RNA-Seq dataset are dropout candidates.
}
\details{
populates a Boolean matrix with the same dimension as nData.
The value is \code{TRUE} for an entry if it
is a dropout candidate; otherwise the value is \code{FALSE}.
}
\examples{
example(cidr)
}

