% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\docType{methods}
\name{scCluster}
\alias{scCluster}
\title{Single-cell Clustering}
\usage{
\S4method{scCluster}{scData}(object, n = NULL, nCluster = NULL,
  nPC = NULL, cMethod = "ward.D2")
}
\arguments{
\item{object}{an scData class object.}

\item{n}{Calinski-Harabasz Index is used to decide which number between 2 and n is optimal as the number of clusters; if \code{NULL} (default), it is set to be nPC*2+2. User should not assign both n and nCluster.}

\item{nCluster}{the number of clusters; if \code{NULL} (default), it is determined automatically. User should not assign both n and nCluster.}

\item{nPC}{the number of PCs used in clustering; by default 4.}

\item{cMethod}{hierarchical clustering method; by default "ward.D2".}
}
\value{
an updated scData class object with the following attributes updated

\item{nCluster}{the number of clusters.}
\item{nPC}{the number of PCs used in clustering.}
\item{clusters}{a vector assigning each cell to a cluster.}
\item{cMethod}{hierarchical clustering method.}
}
\description{
performs heirarchical clustering on \emph{CIDR} principal coordinates.
}
\examples{
example(cidr)
}

