% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIDR.R
\docType{methods}
\name{scDissim}
\alias{scDissim}
\title{CIDR Dissimilarity Matrix}
\usage{
\S4method{scDissim}{scData}(object, correction = FALSE)
}
\arguments{
\item{object}{the scData class object}

\item{correction}{boolean; if \code{TRUE} apply Cailliez correction}
}
\value{
an updated scData class object with the following attribute updated

\item{dissim}{the \emph{CIDR} dissimilarity matrix}
}
\description{
computes the \emph{CIDR} dissimilarity matrix.
}
\examples{
example(cidr)
}

